Refine stream
