+++ example/stream/stream.bak.html
@@ -5,7 +5,7 @@
     <title>Streamgraph</title>
     <script src="../../build/datav.js"></script>
     <script src="../../lib/charts/axis.js"></script>
-    <script src="../../lib/charts/stream.js"></script>
+    <script src="../../lib/charts/stream.bak.js"></script>
     <style type="text/css">
 #chart {
     border-top: 1px dashed #F00;
