+++ lib/charts/hover_line.js
@@ -18,7 +18,7 @@
   HoverLine.prototype.render = function () {
     this.clear();
     var paper = this.owner.paper;
-    var conf = this.owner.defaults;
+    var conf = this.defaults;
     this.indicatorLine = paper.path("M0 0V" + conf.chartHeight).attr({
       stroke: "none",
       "stroke-width": 1,
@@ -38,25 +38,27 @@
     this.highlightLine.attr({"stroke": "white"});
   };
 
-  HoverLine.prototype.refresh = function (xIdx, pathIndex) {
+  HoverLine.prototype.refresh = function (columnIndex, rowIndex) {
     //refresh lines' position
     var owner = this.owner;
     var pathSource = owner.pathSource;
-    var lineX = owner.defaults.chartWidth * xIdx / (owner.source[0].length - 1);
-    var pathSourceCell = pathSource[pathSource.length - 1][xIdx];
+    var lineX = this.defaults.chartWidth * columnIndex / (owner.columnCount - 1);
+    var pathSourceCell = pathSource[pathSource.length - 1][columnIndex];
     this.indicatorLine.attr({
-      path: "M" + lineX + " " + (pathSourceCell.y0 - pathSourceCell.y) + "V" + pathSource[0][xIdx].y0
+      path: "M" + lineX + " " + (pathSourceCell.y0 - pathSourceCell.y) + "V" + pathSource[0][columnIndex].y0
     });
 
-    pathSourceCell = pathSource[pathIndex][xIdx];
-    this.highlightLine.attr({
-      path: "M" + lineX + " " + (pathSourceCell.y0 - pathSourceCell.y) + "V" + pathSourceCell.y0
-    });
-  
-    if (pathIndex === 0 && owner.getDisplayRowInfo(pathIndex).rowIndex === -1) {
-      this.highlightLine.attr({"cursor": "pointer"});
-    } else {
-      this.highlightLine.attr({"cursor": "auto"});
+    if (typeof rowIndex !== 'undefined') {
+      pathSourceCell = pathSource[rowIndex][columnIndex];
+      this.highlightLine.attr({
+        path: "M" + lineX + " " + (pathSourceCell.y0 - pathSourceCell.y) + "V" + pathSourceCell.y0
+      });
+
+      if (rowIndex === 0) {
+        this.highlightLine.attr({"cursor": "pointer"});
+      } else {
+        this.highlightLine.attr({"cursor": "auto"});
+      }
     }
   };
