+++ lib/charts/legend.js
@@ -63,16 +63,15 @@
     this.init();
     this.clear();
     this.legends = [];
-    var owner = this.owner;
     var colorFunc = this.defaults.colorFunc,
       hoverIn = function (e) {
         var index = e.data.index;
-        owner.fire('hoverIn', index);
+        that.fire('hoverIn', index);
         this.highlight(index);
       },
       hoverOut = function (e) {
         var index = e.data.index;
-        owner.fire('hoverOut', index);
+        that.fire('hoverOut', index);
         this.lowlight(index);
       };
 
@@ -104,10 +103,14 @@
   };
 
   Legend.prototype.highlight = function (index) {
-    this.legends[index].css({"background": "#dddddd"});
+    if (typeof index !== 'undefined') {
+      this.legends[index].css({"background": "#dddddd"});
+    }
   };
   Legend.prototype.lowlight = function (index) {
-    this.legends[index].css({"background": "white"});
+    if (typeof index !== 'undefined') {
+      this.legends[index].css({"background": "white"});
+    }
   };
   Legend.prototype.clear = function () {
     this.legend.empty();
