+++ lib/charts/navi.js
@@ -21,16 +21,14 @@
     this.naviBackWidth = 80;
     var conf = this.defaults;
     this.node.css({
-      "border-top": "1px solid #ddd",
-      "border-bottom": "1px solid #ddd",
-      "padding-top": "5px",
-      "padding-bottom": "10px",
-      "padding-left": "10px",
-      "padding-right": "10px",
-      "font": (conf.fontSize + 1) + "px 宋体"
+      "borderTop": "1px solid #ddd",
+      "borderBottom": "1px solid #ddd",
+      "padding": "5px 10px 10px 10px",
+      "fontSize": conf.fontSize + 1,
+      "fontFamily": "宋体"
     });
     this.naviTrace = $("<div></div>").css({
-      "width": conf.legendBesidesWidth - this.naviBackWidth - 50 + "px",
+      "width": conf.legendBesidesWidth - this.naviBackWidth - 50,
       "margin-top": "5px"
     });
