+++ lib/charts/stream_axis.js
@@ -29,6 +29,7 @@
     map = this.map(map);
     this.grouped = _.groupBy(source, map.x);
     this.axis = _.keys(this.grouped);
+    this.range = [0, this.axis.length - 1];
   };
 
   Axis.prototype.init = function () {
@@ -47,7 +48,7 @@
     //all date strings' format are same, string length are same 
     var conf = this.defaults,
       that = this;
-      getPopPath = function (El) {
+    var getPopPath = function (El) {
         //down pop
         var x = 0,
           y = 0,
@@ -92,18 +93,18 @@
         return that.axis[d] || "";
       })(this.paper);
 
-    // this.axisPopText = this.paper.text(0, 11, '')
-    //   .attr({
-    //     "text-anchor": "middle",
-    //     "fill": "#fff",
-    //     "transform": "t" + left + ",0"
-    //   }).hide();
-    // this.axisPopBubble = this.paper.path(getPopPath(this.axisPopText))
-    //   .attr({
-    //     "fill": "#000",
-    //     "transform": "t" + (-10000) + ",0"
-    //   }).toBack()
-    //   .hide();
+    this.axisPopText = this.paper.text(0, 11, this.axis[0])
+      .attr({
+        "text-anchor": "middle",
+        "fill": "#fff",
+        "transform": "t" + left + ",0"
+      }).hide();
+    this.axisPopBubble = this.paper.path(getPopPath(this.axisPopText))
+      .attr({
+        "fill": "#000",
+        "transform": "t" + (-10000) + ",0"
+      }).toBack()
+      .hide();
   };
   Axis.prototype.hideTab = function () {
     this.axisPopText.hide();
@@ -113,9 +114,12 @@
     this.axisPopText.show();
     this.axisPopBubble.show();
   };
-  Axis.prototype.refreshTab = function (timeText, transX) {
+  Axis.prototype.refreshTab = function (index) {
+    var conf = this.defaults;
+    var x = conf.chartWidth * index / (this.axis.length - 1);
+    var transX = x + this.defaults.percentageWidth;
     this.axisPopText.attr({
-      "text": timeText
+      "text": this.axis[index + this.range[0]]
     }).transform("t" + transX + ",0");
     this.axisPopBubble.transform("t" + transX + ",0");
   };
