+++ lib/components/stream.js
@@ -41,6 +41,7 @@
       this.defaults.rowNameUsed = "auto";
       this.defaults.pathLabel = true;
       this.defaults.fontSize = 12;
+      this.defaults.colorCount = 20;
       //this.defaults.axisTickNumber = 8; // axis ticks number
 
       this.defaults.indexMargin = 3; // if dates.length < indexMargin * 2 + 1, do not show label
@@ -60,6 +61,8 @@
 
       this.defaults.legendBesidesWidth = undefined;
       this.defaults.legendBesidesHeight = undefined;
+
+      this.defaults.level = 0;
 
       this.defaults.chartWidth = undefined;//depends on width, do not recommend to change
       this.defaults.chartHeight = undefined;// depends on height, do not recommend to change
@@ -117,12 +120,17 @@
       });
     };
     that.on('changeLevelTo', function (level) {
-      that.defaults.moreConfig.level = level;
+      that.defaults.level = level;
       getBack(that.defaults.moreConfig.level);
     });
     that.on('back', function () {
-      that.defaults.moreConfig.level = that.defaults.moreConfig.level - 1;
-      getBack(that.defaults.moreConfig.level);
+      that.defaults.level = that.defaults.level - 1;
+      getBack(that.defaults.level);
+    });
+    that.legend.on('hoverIn', function (index) {
+      that.stream.highlight(index);
+    }).on('hoverOut', function (index) {
+      that.stream.lowlight(index);
     });
     that.stream.on('enter', function () {
       that.axis.showTab();
@@ -140,8 +148,7 @@
       that.legend.highlight(cur);
       that.tip.setContent(cur);
       //axis pop bubble
-      var lineX = that.defaults.chartWidth * xIdx / (that.chart.source[0].length - 1);
-      that.axis.refreshTab(that.date[xIdx + that.timeRange[0]], lineX + that.defaults.percentageWidth);
+      that.axis.refreshTab(xIdx);
     });
   };
