+++ lib/datav.js
@@ -525,6 +525,7 @@
         ret.hasField = _.any(ret, function (val) {
             return typeof val === 'string';
         });
+        this.mapping = ret;
         return ret;
     };
